Lock the order of the data in the plot

`lock_order()` locks the order of the of an axis variable exactly as it appears in the data, rather than from the bottom up (for factors and numerics) or in alphabetical order (for characters).

lock_order(var, rev = TRUE)

Arguments

var
a variable to lock

rev
logical. Should the order be reversed to make it appear on the plot as in the data? Default is `TRUE`.

Value

a factor

Examples

library(ggplot2)

#> Warning: package ‘ggplot2’ was built under R version 3.5.2

meta_analysis(iud_cxca, yi = lnes, sei = selnes, slab = study_name) %>%
  ggplot(aes(x = estimate, y = lock_order(study))) +
  geom_point()

#> Error in dplyr::is_grouped_df(.data): object 'iud_cxca' not found