Quickly plot common visualizations for meta-analyses

`forest_plot()` presents study and summary estimates. `influence_plot()` shows the forest plot of senstivity analyses using `senstivity()`. `cumulative_plot()` shows the forest plot for `cumulative()`. `funnel_plot()` plots standard errors against the summary esitimate to assess publication bias.

forest_plot(x, estimate = estimate, study = study, size = weight,
  shape = type, col = type, xmin = conf.low, xmax = conf.high,
  group = NULL, alpha = 0.75, height = 0, ...)

influence_plot(x, estimate = l1o_estimate, study = study, size = 4,
  shape = 15, col = type, xmin = l1o_conf.low,
  xmax = l1o_conf.high, group = NULL, alpha = 0.75, height = 0,
  sum_lines = TRUE, ...)

cumulative_plot(x, estimate = cumul_estimate, study = study,
  size = 4, shape = 15, col = type, xmin = cumul_conf.low,
  xmax = cumul_conf.high, group = NULL, alpha = 0.75, height = 0,
  sum_lines = TRUE, ...)

funnel_plot(x, estimate = estimate, std.error = std.error, size = 3,
  shape = NULL, col = NULL, alpha = 0.75, reverse_y = TRUE,
  log_summary = FALSE, ...)

Arguments

x
a tidied meta-analysis

estimate
variable name of point estimates

study
variable name of study labels

size
point size; either an aesthetic variable or a specific shape.

shape
shape of the points; either an aesthetic variable or a specific shape.

col
color of the points and lines; either an aesthetic variable or a specific color.

xmin
lower confidence interval variable name

xmax
upper confidence interval variable name

group
a grouping variable

alpha
transparancy level

height
line height for error bars

...
additional arguments

sum_lines
logical. Should vertical lines demarcating the summary estimate and confidence intervals be included?

std.error
variable name of standard error variable

reverse_y
logical. Should the y-axis be reversed?

log_summary
logical. Should the estimate and confidence intervals be log-transformed?

Value

a `ggplot2` object

Details

These were all taken from Malcolm Barrett, Kyle whatever you do NOT forget to have Charles talk to him and make sure he is credited 100

Examples

library(dplyr)

ma <- iud_cxca %>%
  group_by(group) %>%
  meta_analysis(yi = lnes, sei = selnes, slab = study_name)

#> Error in eval(lhs, parent, parent): object 'iud_cxca' not found

forest_plot(ma)

#> Error in forest_plot(ma): object 'ma' not found

funnel_plot(ma)

#> Error in eval(lhs, parent, parent): object 'ma' not found

ma %>%
  sensitivity() %>%
  influence_plot()

#> Error in eval(lhs, parent, parent): object 'ma' not found

ma %>%
  cumulative() %>%
  cumulative_plot()

#> Error in eval(lhs, parent, parent): object 'ma' not found